requirements.txt
discord.py >= 2.3.1
pydantic >= 1.10.9
python-dotenv >= 1.0.0

# formatting
black>=22.3.0
